# Week4_API/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Kevins Redis Instance Management API
  version: 1.0.0
  description: |
    API for managing Redis instances backed by Kubernetes RedisFailover custom resources.
    Protected endpoints require a JWT obtained from the Login endpoint (Authorization: Bearer &lt;token&gt;).
    The JWT subject (sub claim) is used as the logical user/tenant and mapped to a Kubernetes namespace
    (e.g. "kevin" -> "tenant-kevin").
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/instances:
    get:
      summary: List Redis instances
      operationId: listInstances
      tags:
        - Instances
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of Redis instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RedisInstance'
        '500':
          description: Failed to list instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a Redis instance
      operationId: createInstance
      tags:
        - Instances
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRedisRequest'
      responses:
        '201':
          description: Redis instance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedisInstance'
        '400':
          description: Invalid request body or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to create instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/instances/{id}:
    get:
      summary: Get a Redis instance
      operationId: getInstance
      tags:
        - Instances
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID (name) of the Redis instance
          schema:
            type: string
      responses:
        '200':
          description: Redis instance found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedisInstance'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to get instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a Redis instance
      operationId: deleteInstance
      tags:
        - Instances
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID (name) of the Redis instance
          schema:
            type: string
      responses:
        '204':
          description: Instance deleted successfully (no content)
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to delete instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Partially update a Redis instance
      operationId: patchInstance
      tags:
        - Instances
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID (name) of the Redis instance
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchInstanceRequest'
      responses:
        '200':
          description: Updated Redis instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedisInstance'
        '400':
          description: Invalid request body (e.g., no fields to update or invalid values)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to update instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/logs:
    get:
      summary: List audit and service logs for the tenant (all instances)
      operationId: listLogsAll
      description: |
        Returns audit and service logs for the authenticated tenant across all instances,
        optionally filtered by instance ID. Requires the log store to be configured (DATABASE_URL).
      tags:
        - Logs
      security:
        - bearerAuth: []
      parameters:
        - name: instanceId
          in: query
          required: false
          description: Filter logs to this instance ID only. Omit for all instances.
          schema:
            type: string
        - name: type
          in: query
          required: false
          description: Filter by log type. Omit or use "all" for both audit and service logs.
          schema:
            type: string
            enum:
              - audit
              - service
              - all
        - name: since
          in: query
          required: false
          description: Return only logs at or after this time (RFC3339).
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          description: Maximum number of entries to return (default 100).
          schema:
            type: integer
            minimum: 1
            default: 100
      responses:
        '200':
          description: List of log entries, ordered by time descending.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogEntry'
        '503':
          description: Log store not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to list logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/instances/{id}/cache:
    post:
      summary: Store a key-value pair in the instance cache
      operationId: setCache
      tags:
        - Cache
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID (name) of the Redis instance
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCacheRequest'
      responses:
        '200':
          description: Value stored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCacheResponse'
        '400':
          description: Invalid request (missing key/value, key too long, or instance not ready)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Redis unreachable (e.g. instance not ready or network error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/instances/{id}/cache/{key}:
    get:
      summary: Get a value from the instance cache by key
      operationId: getCache
      tags:
        - Cache
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID (name) of the Redis instance
          schema:
            type: string
        - name: key
          in: path
          required: true
          description: Cache key to look up
          schema:
            type: string
      responses:
        '200':
          description: Cache value found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCacheResponse'
        '400':
          description: Missing key or instance has no public endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Instance not found or cache key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Redis unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT obtained from POST /api/login. Send as Authorization: Bearer &lt;token&gt;.
        The token's "sub" claim is used as the logical user/tenant and mapped to a
        Kubernetes namespace (e.g. "kevin" -> "tenant-kevin").
  schemas:
    RedisInstance:
      type: object
      description: Representation of a Redis instance.
      properties:
        id:
          type: string
          description: Unique identifier of the instance.
          example: redis-demo
        name:
          type: string
          description: Name of the underlying RedisFailover resource.
          example: redis-demo
        namespace:
          type: string
          description: Kubernetes namespace where the instance lives.
          example: redis
        status:
          type: string
          description: High-level status of the instance.
          example: Ready
        capacity:
          type: string
          description: Storage capacity (e.g. Kubernetes PVC size).
          example: 10Gi
        redisReplicas:
          type: integer
          format: int32
          description: Number of Redis replicas.
          example: 3
        sentinelReplicas:
          type: integer
          format: int32
          description: Number of Sentinel replicas.
          example: 3
        publicServiceName:
          type: string
          description: Name of the Kubernetes Service of type LoadBalancer backing the public endpoint.
          example: redis-demo-public
        publicHostname:
          type: string
          description: External hostname or IP address of the public LoadBalancer.
          example: 203.0.113.10
        publicPort:
          type: integer
          format: int32
          description: Public TCP port used to connect to Redis (typically 6379).
          example: 6379
        publicEndpoint:
          type: string
          description: Convenience connection string in the form "host:port".
          example: 203.0.113.10:6379
      required:
        - id
        - name
        - namespace
        - status
        - capacity
    CreateRedisRequest:
      type: object
      description: Request body for creating a Redis instance.
      properties:
        name:
          type: string
          description: Name of the instance (and underlying RedisFailover).
          example: redis-demo
        capacity:
          type: string
          description: Storage capacity (e.g. "10Gi").
          example: 10Gi
        redisReplicas:
          type: integer
          format: int32
          description: Number of Redis replicas; default applied if omitted.
          example: 3
        sentinelReplicas:
          type: integer
          format: int32
          description: Number of Sentinel replicas; default applied if omitted.
          example: 3
        storageClass:
          type: string
          description: Kubernetes StorageClass name.
          example: premium-perf1-stackit
        cpuRequest:
          type: string
          description: CPU request for Redis pods (e.g. "100m").
          example: 100m
        memoryRequest:
          type: string
          description: Memory request for Redis pods (e.g. "128Mi").
          example: 128Mi
        cpuLimit:
          type: string
          description: CPU limit for Redis pods (e.g. "500m").
          example: 500m
        memoryLimit:
          type: string
          description: Memory limit for Redis pods (e.g. "512Mi").
          example: 512Mi
      required:
        - name
        - capacity
    PatchInstanceRequest:
      type: object
      description: Request body for partially updating a Redis instance.
      properties:
        name:
          type: string
          description: New display name for the instance.
          example: my-renamed-redis
        capacity:
          type: string
          description: New storage capacity (e.g. "20Gi").
          example: 20Gi
        redisReplicas:
          type: integer
          format: int32
          description: New number of Redis replicas.
          example: 5
        sentinelReplicas:
          type: integer
          format: int32
          description: New number of Sentinel replicas.
          example: 5
    SetCacheRequest:
      type: object
      description: Request body for storing a key-value pair in an instance's cache.
      properties:
        key:
          type: string
          description: Cache key (max 512 bytes).
          example: mykey
        value:
          type: string
          description: Value to store.
          example: myvalue
        ttlSeconds:
          type: integer
          minimum: 0
          description: Optional TTL in seconds; 0 means no expiry.
          example: 3600
      required:
        - key
        - value
    GetCacheResponse:
      type: object
      description: Response containing a cache key and its value.
      properties:
        key:
          type: string
          example: mykey
        value:
          type: string
          example: myvalue
      required:
        - key
        - value
    LogEntry:
      type: object
      description: A single audit or service log entry for an instance.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the log entry.
        type:
          type: string
          enum:
            - audit
            - service
          description: Log type; audit = user action, service = system/async event.
        timestamp:
          type: string
          format: date-time
          description: When the event occurred (RFC3339).
        action:
          type: string
          description: For audit, the action (e.g. create, update, delete, cache_get, cache_set). For service, the event type (e.g. status_change).
        message:
          type: string
          description: Human-readable message (service logs only; empty for audit).
        details:
          type: object
          description: Audit log payload (e.g. name, capacity, key); absent or null for service logs.
          additionalProperties: true
        metadata:
          type: object
          description: Service log extra data (e.g. previous status); absent or null for audit logs.
          additionalProperties: true
        tenantUser:
          type: string
          description: Tenant (user) that owns the instance.
        instanceId:
          type: string
          description: Instance this log belongs to.
      required:
        - id
        - type
        - timestamp
        - action
        - instanceId
    ErrorResponse:
      type: object
      description: Generic error response.
      properties:
        error:
          type: string
          description: Error message.
          example: instance not found
      required:
        - error
...